smartrecipe/requirements.txt
altgraph==0.10.2
amqp==1.4.6
anyjson==0.3.3
asn1crypto==0.24.0
bdist-mpkg==0.5.0
billiard==3.3.0.20
#bonjour-py==0.3
celery==3.1.18
certifi==2019.9.11
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
data-importer==2.2.1
django>=1.11.19
django-admin-tools==0.7.1
django-extensions==2.1.5
django-favicon==0.1.1
django-filter==1.0.0
django-suit==0.2.15
djangorestframework==3.9.1
#edgegrid-python==1.0.10
enum34==1.1.6
future==0.17.1
gitdb==0.6.4
GitPython==2.0.2
graphviz==0.4.2
idna==2.8
ipaddress==1.0.22
jdcal==1.4.1
kombu==3.0.26
macholib==1.5.1
Markdown==2.6.7
#matplotlib==1.3.1
modulegraph==0.10.4
ndg-httpsclient==0.4.0
numpy==1.16.5
openpyxl==2.2.3
ordereddict==1.1
pbr==1.8.1
Pillow==6.2.0
py2app==0.7.3
pyasn1==0.1.9
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests>=2.20.0
scipy==1.2.0
simplejson==3.6.5
six==1.12.0
smmap==0.9.0
typing==3.6.6
urllib3==1.25.7
#xattr==0.6.4